feat: if user didnt complete 3 step in sign-up dont allow to go to any page, but only signup
