+++ src/app/(auth)/callback/route.ts
@@ -1,4 +1,4 @@
-import {createClient} from "@/utils/superbase/server";
+import {createClient} from "@/utils/supabase/server";
 import {NextResponse} from "next/server";
 // The client you created from the Server-Side Auth instructions
