+++ src/middleware.ts
@@ -1,4 +1,4 @@
-import {updateSession} from "@/utils/superbase/middleware";
+import {updateSession} from "@/utils/supabase/middleware";
 import {type NextRequest} from "next/server";
 
 export async function middleware(request: NextRequest) {
