+++ src/validation/signUpValidation.ts
@@ -18,7 +18,7 @@ export const signUpSchemaStep3 = z.object({
     ),
   username: z
     .string()
-    .min(3, {message: "Username must be at least 3 characters"})
+    .min(4, {message: "Username must be at least 4 characters"})
     .max(20, {message: "Username must be at most 20 characters"})
     .regex(/^[a-zA-Z][a-zA-Z0-9_-]*$/, {
       message:
